refactor: consolidate variant shredding helpers
