+++ crates/sail-common-datafusion/Cargo.toml
@@ -14,6 +14,7 @@ ryu = { workspace = true }
 datafusion = { workspace = true }
 datafusion-common = { workspace = true }
 datafusion-expr = { workspace = true }
+arrow-schema = { workspace = true }
 thiserror = { workspace = true }
 either = { workspace = true }
 lexical-core = { workspace = true }
@@ -25,6 +26,7 @@ prost = { workspace = true }
 tonic = { workspace = true }
 pin-project-lite = { workspace = true }
 parquet-variant-compute = { workspace = true }
+parquet-variant = { workspace = true }
 parquet-variant-json = { workspace = true }
 tokio = { workspace = true }
 serde_arrow = { workspace = true }
