+++ crates/sail-common-datafusion/src/lib.rs
@@ -15,3 +15,4 @@ pub mod streaming;
 pub mod system;
 pub mod udf;
 pub mod utils;
+pub mod variant;
